pyproject.toml
[tool.poetry]
name = "movetobot"
version = "0.1.0"
description = "movetobot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.0"
nonebot2 = "^2.0.0rc4"
parse = "^1.19.0"
numpy = "^1.24.3"
pil-utils = "^0.1.7"
nonebot-adapter-onebot = "^2.2.3"
requests = "^2.30.0"
gputil = "^1.4.0"
nonebot-plugin-localstore = "^0.4.1"
pygithub = "^1.58.2"
nonebot-plugin-apscheduler = "^0.2.0"
bs4 = "^0.0.1"
crypto = "^1.4.1"
cryptography = "^41.0.1"
pycryptodome = "^3.18.0"
rsa = "^4.9"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_localstore", "nonebot_plugin_apscheduler"]
plugin_dirs = ["plugins"]
adapters = [{name = "mirai2", module_name = "nonebot.adapters.mirai2"}]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
